fix: remove internal fields, retrieve deleted item and improve error management
