+++ integration/src/main/scala/io/funkode/portfolio/PortfolioService.scala
@@ -1,0 +1,48 @@
+/*
+ * Copyright (c) 2023 io.funkode
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of
+ * this software and associated documentation files (the "Software"), to deal in
+ * the Software without restriction, including without limitation the rights to
+ * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
+ * the Software, and to permit persons to whom the Software is furnished to do so,
+ * subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+ * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+ * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+ * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+ * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+package io.funkode.portfolio
+
+import zio.*
+
+import io.funkode.portfolio.model.*
+
+type PortfolioCall[R] = IO[PortfolioError, R]
+
+trait PortfolioService:
+
+  def addNetwork(network: Network): PortfolioCall[Network]
+  def addTransaction(transaction: Transaction): PortfolioCall[Transaction]
+  def networkTransactions(network: Network): PortfolioCall[Iterator[Transaction]]
+/*
+object PortfolioService:
+
+  class Impl[Enc[_], Dec[_], Body](resourceStore: ResourceStore[Enc, Dec, Body]) extends PortfolioService:
+
+    def addNetwork(network: Network): PortfolioCall[Network] =
+      resourceStore.store(network)
+
+    def addTransaction(transaction: Transaction): PortfolioCall[Transaction] =
+      resourceStore.store(transaction)
+
+    def networkTransactions(network: Network): PortfolioCall[Iterator[Transaction]] =
+      resourceStore.fetchRel(network, "transactions")
+ */
